rename: detect to segment
