+++ src/cala/nodes/segment/catalog.py
